Refactor creating compiled container file
Since Compiler only check if the compiled container exists then loads it
this means in high traffic situations the compiled container file might
exist but has not been written yet leading to an error. I looked into
how twig handles writing its cache and they write to a temp file then
renames it. This approach makes a lot of sens since the compiler only
cares that there are a complete php file to load.

I might rethink the compiler in a later version.  I think creating the
compiled container should propably not be performed in runtime since
multiple processes might do the same compiling which is redundant. A
seperate cache warmer so we only need to check for existing files and
let another process create the compiled container.